kv_agent/pyproject.toml
[project]
name = "kv-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.9.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
